# animals-place-api/animals-place/openapi.yml
openapi: 3.1.0
info:
  title: Animals Place API
  description: API for an Animal Place Meeting app develloped in the ArchiOWeb course
  termsOfService: http://swagger.io/terms/
  contact:
    email: jorane.membrez@heig-vd.ch
  version: 2.0.0
  externalDocs:
    description: Link to the Github repository
    url: https://github.com/JoraneMembrez/archiOWeb-animalsPlace
servers:
  - url: http://localhost:8000
    description: Local server
  - url: https://archioweb-animalsplace.onrender.com/
    description: Heroku server
...
paths:
  /users :
     get :
      tags:
       - users
      summary: Retourne tous les utilisateurs de l'application de manière paginée
      description: Il faut être admin pour pouvoir exécuter cette requête
      parameters:
        - name: page
          in: query
          description: Numéro de page à récupérer
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 10
          description: Le nombre d'utilisateur par page
      responses: 
        "200":
          description: Successful response
          content:
            application/json:
             schema:
               type: object
               properties :
                 users :
                  type: array
                  items :
                    $ref: "#/components/schemas/UserBasic"
               currentPage :
                 type: integer
               totalPages:
                  type: integer
        "403":
          description : You are not authorized to access this resource
          content:
              text/plain:
                example: Forbidden
      security:
        - bearerAuth: []
     post :
        tags:
          - users
        summary: Création d'un nouvel utilisateur
        requestBody:
          required : true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPost"
        responses:
          "201":
            description: Création réussie
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/UserBasic"
          "409" :
            description : L'utilisateur existe déjà
            content:
              text/plain:
                example: La ressource avec l'email john.doe@gmail.com existe déjà
          "400" :
            description : La requête n'est pas valide
            content:
              text/plain:
                example: Le mot de passe, l'email et le prénom sont obligatoires
  /users/{userID} :
    get :
      tags:
       - users
      summary: Obtenir un utilisateur par son ID, vous devez être administrateur ou l'utilisateur lui-même pour exécuter cette requête.
      parameters:
        - in: path
          name: userID
          schema:
           type: string
           required: true
           example: 6375ef6d1d458a3858c3453e
          description: L'ID de l'utilisateur qu'on souhaite obtenir
      responses:
        "200":
           description: Opération réussie
           content:
              application/json:
                schema:
                  $ref: "#/components/schemas/UserGet"
     
        "404" :
          description : L'utilisateur n'a pas été trouvé
          content:
              text/plain:
                example: L'utilisateur avec l'ID 6375ef6d1d458a3858c3453e n'a pas été trouvé
        "403" :
          description : Vous n'êtes pas autorisé à accéder à cette ressource
          content:
              text/plain:
                example: Interdit
    patch:
          tags:
            - users
          summary: Mise à jour de l'utilisateur par ID
          description: Mettre à jour un utilisateur par son ID. Les utilisateurs administrateurs peuvent mettre à jour n'importe quel utilisateur ; les utilisateurs ne peuvent mettre à jour que leur propre profil.
          parameters:
          - name: userID
            in: path
            example: 6375ef6d1d458a3858c3453e
            description: L'ID de l'utilisateur à mettre à jour
            required: true
            schema:
              type: string
          requestBody:
            required: true
            content:
             application/json:
              schema:
                $ref: '#/components/schemas/UserBasic'
          responses:
            '200':
              description: Utilisateur mis à jour correctement
              content:
                application/json:
                 schema:
                   $ref: '#/components/schemas/UserBasic'
            '400':
              description: L'utilisateur n'a pas été trouvé ou la requête n'est pas valide
              content:
                text/plain:
                  example: L'utilisateur avec l'ID 6375ef6d1d458a3858c3453e n'exite pas
            '403':
              description: Vous n'êtes pas autorisé à accéder à cette ressource
              content:
                text/plain:
                  example: Forbidden
            '401' :
              description : Vous n'êtes pas autorisé à accéder à cette ressource
              content:
                text/plain:
                  example: Unauthorized
          
    delete:
        tags :
          - users
        summary: Supprime un utilisateur par son ID
        description: Supprimer un utilisateur par son ID. Les administrateurs peuvent supprimer n'importe quel utilisateur ; les utilisateurs ne peuvent supprimer que leur propre profil.
        parameters:
          - name: userID
            in: path
            description: L'ID de l'utilisateur a supprimer
            example: 6375ef6d1d458a3858c3453e
            required: true
            schema:
              type: string
        responses:
          '200':
            description: Utilisateur supprimée avec succès
            content : 
              text/plain:
                example: Utilisateur avec l'ID 6375ef6d1d458a3858c3453e supprimé avec succès
          '404':
            description: L'utilisateur n'a pas été trouvé
            content:
              text/plain:
                example: L'utilisateur avec l'ID 6375ef6d1d458a3858c3453e n'existe pas
          '403':
            description: Vous n'êtes pas autorisé à accéder à cette ressource
            content:
              text/plain:
                example: Forbidden
          '401' :
            description : Vous n'êtes pas autorisé à accéder à cette ressource
            content:
              text/plain:
                example: Unauthorized
  /users/matchedUsers : 
    get :
        tags:
          - users
        summary: Obtenir tous les utilisateurs qui ont matché avec l'utilisateur logé
        responses:
          "200":
            description: Opération réussie
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: "#/components/schemas/UserBasic"
          "401":
            description : "Vous n'êtes pas autorisé à accéder à cette ressource"
            content :
              text/plain:
                example: Unauthorized       
        security:
          - bearerAuth: []
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, for documentation purposes only
  schemas :
    UserBasic:
      type: object
      properties:
        _id:
          type: string
          example: 6375ef6d1d458a3858c3453e
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          example: john.doe@gmail.com
        password:
          type: string
          example: "1234"
        registrationDate:
          type: string
          format: date-time
          example: 2021-04-13T14:30:00.000Z
        address :
          type: string
          example: "Rue de la gare 12, 2800 Delémont"
        animals :
          type: Object
          $ref : "#/components/schemas/Animal"
        role:
          type: string
          enum:
           - user
           - admin
          default: user
        location:
          type:
            type: string
            required: true
            enum:
              - Point
            default: Point
          coordinates:
            type:
             - number
            required: true
            validate:
              validator: validateGeoJsonCoordinates
              message: "{VALUE} is not a valid longitude/latitude(/altitude) coordinates array"
          default:
            - 6.647778558579233
            - 46.78060279685718 
    UserPost :
      allOf :
        - type : object
          properties :
            firstName: 
                type : string
                example : John
            email :
              type : string
              example : john.doe@gmail.com
            password :
              type : string
              example : 1234
    UserGet :
      allOf :
        - $ref : "#/components/schemas/UserBasic"
        - type : object
          properties :
            _id :
              type : string
              example : 6375ef6d1d458a3858c3453e
            registrationDate :
              type : string
              format : date-time
              example : 2021-04-13T14:30:00.000Z
    Animal :
      allOf :
        - type : object
          properties :
            _id:
              type: string
              example: 6375ef6d1d458a3858c3453e
            name : 
              type : string
              example : "Bobby"
            species :
              enum : 
                - chien
                - chat
                - lapin
                - furet
                - hamster
                - oiseau
                - tortue
                - poisson
                - souris
                - camélémon
                - serpent
                - araigné
                - autre
              example : chien
            age :
              type : integer
              example: 5
            gender :
              enum :
                - Male
                - Female
                - Other
            profilePictureUrl :
              type : string
            description :
                type : string
                example : "Bobby like painting and eating"
            picturesUrl :
                type : array
                items :
                  type : string
                example : [https://www.google.com/url?sa=i&url=https%3A%2F%2Fwww.woopets.fr%2Fchien%2F]
            location :
                type : String,
                example : "Rue de la gare 12, 2800 Delémont"
            favoriteActivites :
                type : array
                items :
                  type : string
                example : ["painting", "eating"]
            matches :
                type : array
                items :
                  $ref : "#/components/schemas/Animal"
                example : ["6375ef6d1d458a3858c3453e"]
            createdDate :
                type : string
                format : date-time
                example : 2021-04-13T14:30:00.000Z
            owner :
                $ref : "#/components/schemas/UserBasic"
                example : "6375ef6d1d458a3858c3453e"
            animal_liked :
                type : array
                items :
                  $ref : "#/components/schemas/Animal"
                example : ["6375ef6d1d458a3858c3453e"]
...